backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1
Pillow==10.0.0
PyJWT==2.10.1
APScheduler==3.10.4
